Fix AutomaticAnalyticsLoggerTest
Summary:
`AutomaticAnalyticsLoggerTest` has a strange behavior: when the test class is run as a whole, the tests pass. However, if some of its test cases run independently, the tests fail. It's caused because the test tries to verify whether `InternalAppEventsLogger` is created during logging. However, at the initialization of `AutomaticAnalyticsLogger` an instance of `InternalAppEventsLogger` will be created.

This diff fixes the bug by verifying the actual logging behavior rather than object creation.

Reviewed By: jingping2015

Differential Revision: D28916089

fbshipit-source-id: bd5619ec4ab11eb61edd0a1eb94054a1f43606f1