daemon: Send ACTION_VIEW intents as SDK messages to client
Summary:
Design doc for opening links on client: https://fb.quip.com/RBv9Am0RhnIq

This diff adds an activity to the Android CG Daemon to catch ACTION_VIEW intents representing links and send an SDK message containing the URL to the worker.
This will then be used by D27619993 to open a dialog and open the link on the client side.

The response from JS is currently ignored like with the message to open the Play Store because we do not require information to be sent back.

Differential Revision: D28324489

fbshipit-source-id: 6aadfd6b466ecfc81e8fd830c7832ebd8bfb55fb