Make AppEventStore internal and migrate related tests
Summary:
Due to the visibility problem in kotlin, we expose a lot of internal classes/methods to allow unit tests to access them. By moving the tests related to an internal class to the kit's own test package, we can add the `internal` modifer back to protect the visibility.

It's harder than we thought before. Many tests classes are coupled together. In order to verify that we are ready for doing this refactor, I do an experiment with `AppEventStore` and fix the configuration for building/executing the tests inside the CoreKit.

Reviewed By: jingping2015

Differential Revision: D29345011

fbshipit-source-id: 452ba959ec7c2f36baaecc49031a5b584de7b2b9