Deprecate bolts.Capture and wait for its removal in next major release
Summary:
As title. Migrate `CancellationTest` to Kotlin to avoid using Capture

`Capture` is used to capture a variable in the context for creating a closure. These capturing is supported by Kotlin by default, so we should not use it anymore. I deprecate it because some client code may (highly unlikely) still use it.

Reviewed By: jingping2015

Differential Revision: D29719597

fbshipit-source-id: 52518826c9da7be5e6f08fde4cad75f1c97cc3fe