Address the flakyness in TaskTest.testDelay
Summary:
The old `testDelay` relies on `Thread.sleep` to wait for the completion of the delayed task. However, it's not reliable since any executors can only promise the command to be run "after" the delay, but not "at" the delay.

Since bolts library should delegate these kind of "waiting" to `BoltsExecutors.schedule()`, I add another command to the executor so that when assertions are executed, we can make sure that all commands within the time period are finished.

Reviewed By: jingping2015

Differential Revision: D30793311

fbshipit-source-id: 89e7ecb20dad6cd8892594f241ed1e0a5dbd0ee5