App Switch/m-site Handling for ShareDialog Create Tournament
Summary:
This adds the functionality to app-switch to the facebook instead of showing the share dialog overlay. If the facebook app is not installed this also allows for the m-site version of ShareDialog.

Todo:
Return Tournament_id to the user ainsley
Returning to game when cancelling from server side ainsleyh
Add Unit tests

Reviewed By: ainsleyh

Differential Revision: D31664881

fbshipit-source-id: efee841243b43f9b53178e0260c067702e267aab