Add AppSwitch/m-site handling for ShareDialog Update Tournament
Summary: This adds functionality to update a tournament using the share dialog. Developers pass in a tournament and a score. We will app-switch if the fb4a is installed but overlay is not working. We will show msite if fb4a is not installed. Once the dialog is shown and the user hits share we will update their score in the tournament and share the tournament on their newsfeed.

Differential Revision: D31664882

fbshipit-source-id: f5c8f792f51fddd71195e4b982c2d6e827209dfc