Adds ShareDialog Update Tournament API Overlay
Summary: This adds the social api update an existing tournament and share using the share dialog. The API will open the share dialog using android overlay if the fb4a app is installed. Once the user hits share we will use the tournament ID from the provided tournament to look up said tournament and update the users score. After updating the score we will share it to the user's timeline.

Reviewed By: ainsleyh

Differential Revision: D31741128

fbshipit-source-id: 00959019329dd1b149aa74944e0200642b572d93