Adds ShareDialog Create Tournament API Overlay
Summary:
This adds the social api create tournament using the sharedialog. The API will open the share dialog using android overlay if the fb4a app is installed. Once the user hits share we will use the config data the developer provided to create the tournament and share it on the user's newsfeed.

TODO:
Returning the tournament_id of the created Tournament ainsleyh
Unit tests

Reviewed By: ainsleyh

Differential Revision: D31664884

fbshipit-source-id: 08ad7b11a261d7be9a08c7b889b9af5d1571b49c