Main (#1021)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Enable 'needs-attention' github action. This action uses the labels 'needs-triage' and 'needs-feedback' to indicate the state of the issue. 'needs-feedback' is used when more information is required to start the resolution. 'needs-triage' indicates that a repo administrator has to triage the issue. This pull request also prevents stale action to close issues tagged with 'needs-triage' and 'acknowledged' labels.

Pull Request resolved: https://github.com/facebook/facebook-android-sdk/pull/1021

Reviewed By: linmx0130

Differential Revision: D31897884

Pulled By: xilosada

fbshipit-source-id: 356b59b6f8727a9632117566f8a513b07321dae3