Refactor AccessTokenManager to remove construction dependency on GraphRequest
Summary: As title. This diff refactors to avoid calling the constructors of `GraphRequest` directly. By using the existing `newGraphPathRequest`, we can create the exactly same requests without adding a new factory method.  Then it allows unit test coverage data to be computed correctly and the unit tests are easier to maintain.

Reviewed By: panavenue

Differential Revision: D32925457

fbshipit-source-id: 88731d2a8040ef9626b33b7c0cebd8ffef6be943