Extract construction of InternalAppEventsLogger
Summary:
TL;DR: in order to get accurate test coverage data, we need to avoid `PowerMockito.whenNew` to mock the construction of an object. This diff replaced the constructions of InternalAppEventsLogger with new factory methods.

PowerMockito is **not compatible** with Jacoco, the tool for obtaining test coverage.

The root cause is that PowerMockito will do bytecode level operations to replace the class to be mocked. Generally it's fine because we don't treat the replaced code as part of the "covered code". The problem is how PowerMockito implement `whenNew`. It will manipulate the classes who create the objects, rather than the classes whose objects are created.

For example, we have a class A. It creates an object of class B with B's constructors. We wish to mock the class B. What we need to declare in `PrepareForTest(...)` is A rather than B. The class A will be modified during the runtime.

`InternalAppEventsLogger` is the common class B in our codebase. A lot of classes create a logger instance to implicitly send the log to AppEvents. In order to test the logging behaviors, we have to mock `InternalAppEventsLogger`, which in fact modifies the classes being tested in the runtime. This procedure blocks Jacoco to get correct test coverage data.

By adding a factory function, we can simply mock the companion object of `InternalAppEventsLogger` and return a mock logger during the tests.

Another approach is to do real dependency injection like what we do for iOS: adding a `InternalAppEventsLoggerProvider` to do all these stuffs. It's something we may need to do if we move AppEvents out of the core kit. Since code coverage will be an important metric for our team in the next half, I would like to fix it with an easy approach and do BE stuffs during the lightweight SDKs project is executed.

One reason I hesitate to do real dependency injection in Android SDK is because unlike iOS where we can control whether to include a method (declaration) in the binary with C macros, any setters will be kept in the binary code. There's no any forms of conditional compilation in the whole Java/JVM/Android ecosystems... The result is that we will have a huge bunch of setters merely for testing remaining in the generated bytecode. The tradeoff between testability and binary size is hard.

Reviewed By: jawwad

Differential Revision: D32749218

fbshipit-source-id: 80f3958f17ace9ad018d8a51a12e283e102a8e39