Validate.notNullOrEmpty returns nonnull value
Summary:
We use `notNullOrEmpty` to validate the values from parcel in dealing with access tokens. However, the original `notNullOrEmpty` cannot hint the Kotlin compiler that the value will be nonnull after the validation, so we have to add `checkNotNull` to make the Kotlin compiler happy. This call will add a redundant null check.

By allowing `notNullOrEmpty` to return the nonnull value, we can remove `checkNotNull` following the `notNullOrEmpty`.

Reviewed By: jvlyn

Differential Revision: D33341826

fbshipit-source-id: 4b155f0de426bbdaddc4482f705fa9fd1964dc7d