Fix network IO on main thread
Summary:
As title. By executing `processCodeExchange` on side threads to avoid network IO on main thread exception without introducing timeout/lock or any other waiting on the main thread.

If the `processCodeExchange` cannot return in a short time, the app will keep waiting as if the login procedure is not complete and the spinner in LoginFragment will also keep rotating.

Reviewed By: alsun2001

Differential Revision: D33851822

fbshipit-source-id: 23c176545377ed29a0213a739920f8de84addc80