Remove the dates next to the versions in CHANGELOG.md since they are not accurate
Summary:
The version headings in the changelog contain a date, but this is actually the date that the version was bumped, not the date it was released so this is misleading

For example, for 13.1.0 it shows 2022-03-14 but this was actually released on 2022-03-21

```
## [13.1.0] - 2022-03-14
```

The dates could be updated to reflect the accurate release dates but it would still be difficult to maintain the correct release date going forward. We generally know what the release date will be (i.e. 7 days in advance) but it may sometimes differ.

Also if someone is interested in looking at actual release dates they can easily see it from the releases page: https://github.com/facebook/facebook-android-sdk/releases

Reviewed By: linmx0130

Differential Revision: D35068377

fbshipit-source-id: 732780f71531969133478e8006873e98cb58bfa0