Expose country code information from android daemon
Summary: Adds support for fetching country ISO from android SDK and adds a button to the CG SDK Channel Test App to test this functionality. The daemon returns "ZZ" to indicate no country information is present. If country information is passed in from www on startup then the daemon caches it.

Differential Revision: D35692252

fbshipit-source-id: 93235ed3797b9e2b0144b3631b29323ef77d96d7