Fix back button behavior for QR code on Catalina
Summary:
Modifies the code in AlohaDeviceAuthDialog.java related to animating the QR code when it is clicked on on Catalina to animate it and make it larger. There is an issue where if users hit the back button while the QR code is expanded, instead of returning to the regular split QR view the flow will navigate to the previous screen instead.

(note the user can also dismiss the QR view by clicking the center dpad button again.)

This fix also involved a slight change to DeviceAuthDialog.kt to allow for suppressing the regular back button functionality.

Differential Revision: D37332607

fbshipit-source-id: 846e767e0c53fd775d53daaa209e5fa0b6a2738f