Use buildSrc for dependencies
Summary:
Using `constants.gradle` is an older way of managing dependencies. The recommendation is now to use the kotlin-dsl plugin which allows an reserved `buildSrc` module which is accessible to all other modules. This enables IDE support which is a huge QoL improvement when dealing with dependencies as we no longer have to rely on full text search only.

See:
https://docs.gradle.org/current/userguide/organizing_gradle_projects.html
https://proandroiddev.com/better-dependencies-management-using-buildsrc-kotlin-dsl-eda31cdb81bf
https://github.com/antoniolg/Bandhook-Kotlin

Reviewed By: linmx0130

Differential Revision: D38715544

fbshipit-source-id: a700cf59ce10c1ff863a88b060b886c1b19e61fa