Convert LoginButton.java and DeviceLoginButton.java to Kotlin
Summary:
Converting the files in scope of the effort migrating the codebase from Java to Kotlin.
Converting both files in the same commit as `DeviceLoginButton` extends `LoginButton` and is a small and simple class.

Also, I realized that my IDE was configured to use 4 spaces indentation and all existing files use 2 spaces, so I'm also re-formatting the files that I changed in my recent commits.

Reviewed By: mingcaozhang

Differential Revision: D41032469

fbshipit-source-id: f7e4fa85497559933b36c6b1bce03b9429a15dad