Fix device dialog layout
Summary: The fragment had the device layout hardcoded to 600dp which probably makes some kind of sense if you assume it’s running on a tv or tablet (although even then it may be way too small depending on the screen size). Better to just have it match the bounds of the parent view with a few dps of padding.

Reviewed By: jawwad

Differential Revision: D41270275

fbshipit-source-id: c88133c1ff02e56a9294a1f81b304b789b6b538f