Implement processParametersForProtectedMode() in ProtectedModeManager
Summary:
## Context
Protected Mode is a feature required from Legal. As the first line of protection on Client-side, this feature will drop all the invalid parameters in app events which are sent to Facebook Server to ensure we are not violating any rule against the user's privacy, e.g. health related data.

## Change in this diff
- Implement processParametersForProtectedMode(), it will drop all the non standard parameters
- Add unittest

Reviewed By: KylinChang

Differential Revision: D46658869

fbshipit-source-id: 27d4ddffc99e7520bcb1c9edacf129ea03204618