Fetch MACA rule GK
Summary:
## Context
Protected Mode is a feature required from Legal. As the first line of protection on Client-side, this feature will drop all the invalid parameters in app events which are sent to Facebook Server to ensure we are not violating any rule against the user's privacy, e.g. health related data.

## Change in this diff
- Fetch GK for MACA rule

Reviewed By: KylinChang

Differential Revision:
D46871303

Privacy Context Container: L1196064

fbshipit-source-id: 38d421f893b53fd49142fca71f906700d0484045