Update standard params list in ProtectedModeManager
Summary:
## Context
Protected Mode is a feature required from Legal. As the first line of protection on Client-side, this feature will drop all the invalid parameters in app events which are sent to Facebook Server to ensure we are not violating any rule against the user's privacy, e.g. health related data.

For App Standard Params list, we align with D46547847. We are using OffsiteSignalsStandardDataUtil, AppCustomEventField & AppEventsIgnoredParams to get a list of App standard parameters.

## Change in this diff
- Update the default Standard params in SDK from the source: https://docs.google.com/document/d/1i30GSHcnTqP-ZLF4dD8UXVGxPb7JGs8oKOcmZOjG7ew/edit

Reviewed By: KylinChang

Differential Revision:
D46948169

Privacy Context Container: L1196064

fbshipit-source-id: d912413a524cbdb0932d93ae7cf39a19081be81f