Fetch and parse Protected Mode Rules in FetchedAppSettingsManager
Summary:
## Context
Protected Mode is a feature required from Legal. As the first line of protection on Client-side, this feature will drop all the invalid parameters in app events which are sent to Facebook Server to ensure we are not violating any rule against the user's privacy, e.g. health related data.

## Change in this diff
- parse standard parameters and maca rules in FetchedAppSettingsManager

Reviewed By: KylinChang

Differential Revision:
D46916265

Privacy Context Container: L1196064

fbshipit-source-id: c33939bd433cd8c66a6b3f19ef0e8fac1046812a