Loading MACA rules from AppSettings
Summary:
## Context
Protected Mode is a feature required from Legal. As the first line of protection on Client-side, this feature will drop all the invalid parameters in app events which are sent to Facebook Server to ensure we are not violating any rule against the user's privacy, e.g. health related data.

## Change in this diff
- load MACARules from AppSettings
- If there is no non_empty / non_null MACA rules, MACARuleMatchingManager will **NOT** be enabled.

Reviewed By: KylinChang

Differential Revision:
D46918126

Privacy Context Container: L1196064

fbshipit-source-id: f23c343377c3f8b19b4fd4a52549199b161766b5