Loading Standard params from App settings
Summary:
## Context
Protected Mode is a feature required from Legal. As the first line of protection on Client-side, this feature will drop all the invalid parameters in app events which are sent to Facebook Server to ensure we are not violating any rule against the user's privacy, e.g. health related data.

## Change in this diff
- Load Standard Params from App Settings
if loading fails, we use the default standard params defined inside the ProtectedModeManager class

Reviewed By: KylinChang

Differential Revision:
D46954711

Privacy Context Container: L1196064

fbshipit-source-id: c02906462013181d6a436801bd8d5857b95526cd