3/n Load the blocklist and check if the event is in blocklist
Summary:
## Context
Blocklist Events is a feature introduced from the integrity perspective. As the first line of protection on Client-side, this feature will drop all the app events which are in the blocklist to ensure we are not sending the data that the advertisers don't want or don't use to minimize the risk of storing sensitive user data.

## Change in this diff
- Load the blocklist and store them in the Set
- Check if the event is in the blocklist

Reviewed By: KylinChang

Differential Revision:
D51487237

Privacy Context Container: L1214531

fbshipit-source-id: 952b68b030817a3be77c80850fbde8efcd73baba