2/n Add RedactedEventsManager
Summary:
## Context
Redacted Events is a feature introduced from the integrity perspective. As the first line of protection on Client-side, this feature will replace the app event name which are in the redacted event list to ensure we are not sending the sensitive event name to minimize the risk of storing sensitive user data.

## Change in this diff
- Add new feature FilterRedactedEvents in FBSDK. See the GK check for FBSDK on Server-side: D50993991

Reviewed By: KylinChang

Differential Revision: D51628453

fbshipit-source-id: ec728d56d513ab93e005585cfcaa57a1403bc8c8