3/n Filter redacted events
Summary:
## Context
Redacted Events is a feature introduced from the integrity perspective. As the first line of protection on Client-side, this feature will replace the app event name which are in the redacted event list to ensure we are not sending the sensitive event name to minimize the risk of storing sensitive user data.

## Change in this diff
- Load the redacted events
- Process event redaction

Reviewed By: KylinChang

Differential Revision: D51623298

fbshipit-source-id: 5bd8fc7684a0bc162033fb3f77e9f7064b2d9a2b