Extend app_events_kill_switch for install event
Summary:
**Context**
We will block the prohibited Apps sending the app events to our server on Client side within FBSDK.
The prohibited apps will be identified on the server and return true for `app_events_killswitch` GK if it's prohibited app.

**Change in this diff**
- extend the check for `app_events_killswitch` before publishing install.

(NOTE)**Note:**
This approach will only guarantee that we are checking the real GK value instead of using default one when `activateApp()` is automatically called by SDK. Which is consistent with the behavior of OnDeviceEventProcessing.
Manually calling `activateApp()` will not be guaranteed that the GK values will return before checking the app_events_killswitch GK.**

Reviewed By: KylinChang

Differential Revision: D51858120

fbshipit-source-id: 3eff9f0bf5a5e91760fd1c90c84a363b6b2ab2ec