Update targeted SDKs for Google Play Internal Release
Summary:
In order to test out the in-app purchase functions in the coffeeshop app, we need to deploy it to Google Play.

In order to do that, we need to make some slight modifications to the coffeeshop build to address problems like:
 {F1788184924}

In this diff, I finish making those.

Now, we just need to await Google to confirm my identity (I had to send in a picture of my ID).

Reviewed By: xta0

Differential Revision: D60700415

fbshipit-source-id: fcd135c495159d27c022469a7779199a77ad34f9