Initialize billing client for GPBL v5-v7
Summary: In this diff, we create the billing client for GPBL v5-v7. This is largely the same as it was for v3, but instead of creating an internal class to implement the invocation handler, we let the companion object itself implement the invocation handler (this seems to be cleaner, fewer lines of code, etc.).

Reviewed By: jjiang10

Differential Revision: D60803355

fbshipit-source-id: 0c055ac7293a6f6636b657de9315b1bc29a0afb3