Create queryPurchasesAsync method for GPBL v5-v7
Summary:
In this diff, I call the new queryPurchasesAsync api that is offered in GPBL v5-v7.

In order to write better unit tests, I moved all of the listeners to be handled by one InvocationHandler, ```ListenerWrapper```. Because, in some cases, we want to pass more arguments to the listener than just those that the listener method receives from Google, I have added an optional parameter that is passed into this ListenerWrapper's constructor.

Reviewed By: jjiang10

Differential Revision: D60875746

fbshipit-source-id: 8cd11d66a413010c29e042c02580107bc436c416