Created queryProductDetailsAsync method for GPBL v5-v7
Summary:
One of the critical GPBL methods that has changed since v3 is queryProductDetailsAsync. This is a method that serves to replace the deprecated and removed querySkuDetailsAsync.

In this diff, I implement the logic to call this new method and handle the response the same way that the previous implementation did.

Reviewed By: jjiang10

Differential Revision: D60943780

fbshipit-source-id: 9948b070ed3ddd11dba4685560596d93911623c1