Extract constants from billing client wrappers
Summary:
In both of the billing client wrapper classes, we were defining a long list of constants to refer to the imported classes and methods.

To removed duplicate strings and make the code more readable, I have extracted all of the constants into an organized IAPConstants class. This is also a good way to visualize which classes we are importing for each implementation and where the overlap is.

Reviewed By: jjiang10

Differential Revision: D60975717

fbshipit-source-id: 0881d73ae813adb4e3ec4491e90531a6cb9358a9