Automatically query product details with results of querypurchases/querypurchasehistory
Summary: In order to query product details, we need to pass in a list of product Ids ([docs](https://developer.android.com/reference/com/android/billingclient/api/QueryProductDetailsParams.Builder)). In order to know what those productIds are, we need to have first queried for purchases or purchase history. Therefore, we need to make the call to queryProductDetailsAsync in the listener of queryPurchasesAsync and queryPurchaseHistoryAsync.

Reviewed By: jjiang10

Differential Revision: D60979363

fbshipit-source-id: 13d73a04dfc937658a1f7b60123f182e1421e13f