Remove 'initialized' field from GPBL v2-v4 wrapper
Summary:
We do not need the ```initialized``` field, as the instance itself will not be created if we were unable to fetch the methods/classes we need from the Google Play Billing Library. As a result, we can just check if the instance is null to determine whether it has been initialized.

Likewise, this solution is still thread-safe, as the ```getOrCreateInstance``` method is marked as ```Synchronized```.

Reviewed By: jjiang10

Differential Revision: D61996407

fbshipit-source-id: 607b61cd96d61eee347442945c12a78f88bb26b9