Remove memory leak due to
Summary:
Passing an application context as a parameter to a static field creates a memory leak. Within the billing client wrapper itself, we don't actually need the application context, just the package name. Therefore, we can just pass the package name as a parameter to the singleton.

In the previous diff, we fixed this memory leak for the wrapper that supports GPBL v5 - v7.

In this diff, we fix the memory leak for the wrapper that supports GPBL v2 - v4.

Reviewed By: jjiang10

Differential Revision: D62002797

fbshipit-source-id: ef3ee457a7a25ca9931c16bde0f55cf7f910386b