Add logging statements for GPBL v2-v4 wrapper
Summary:
In order to achieve parity with the IAP billing client wrapper for GPBL v5-v7, we should log when we are unable to create a billing client. This will be one way for advertisers to notice that in-app purchase events are not being auto-logged.

In order to ensure these logs don't pollute the the application-level logs, we will need to implement functionality that ensures we only attempt to create a billing client wrapper once per session.

Reviewed By: jjiang10

Differential Revision: D62004201

fbshipit-source-id: 90f876a4ccc35d43b24443db52a8333ce9ac904e