Only attempt to create billing client wrappers once
Summary: In the case that we are unable to create a billing client wrapper for IAP auto-logging due to our inability to support a GPBL version, we only want to attempt to create the wrapper once per app session. Previously we were doing this each time in the onActivityResume() callback, which is triggered each time an activity is resumed on the application layer. To fix this, and avoid redundant logging, we should use a thread-safe boolean to keep track of previous failed attempts.

Reviewed By: jjiang10

Differential Revision: D62006298

fbshipit-source-id: 18ca9a093fe3e77fa3f32060e3380b9e3e23aa11