Refactor both billing clients
Summary:
In order to clean up the in-app purchase auto-logging code and adhere to best coding practices, we should create an interface that both billing client wrappers implement. We will call this interface ```InAppPurchaseBillingClientWrapper```. Because the wrapper that supports GPBL v2-v4 is already titled that, we should rename the classes:
1. InAppPurchaseBillingClientWrapperV2V4.kt
2. InAppPurchaseBillingClientWrapperV5V7.kt

to avoid any future confusion.

In the next diff, we will add this interface, ensure both wrappers implement it, and, in the ```InAppPurchaseAutoLogger```, we will use a generic billing client wrapper object to query for in-app purchases.

Reviewed By: jjiang10

Differential Revision: D62034587

fbshipit-source-id: 287dd9b489d598a35f8f3583a6e873006d340837