Create GPBC Wrapper Interface
Summary:
In order to clean up the in-app purchase auto-logging code and adhere to best coding practices, we should create an interface that both billing client wrappers implement.

We can now use this interface to call a generic billing client in ```InAppPurchaseAutoLogger``` or anywhere else in the FBSDK for that matter.

Reviewed By: jjiang10

Differential Revision: D62037255

fbshipit-source-id: 126e57dec5f2d9156f77d9b09678a5c2d2ed4ba0