Support querying subscriptions in GPBL v5-v7
Summary:
In this diff, we support the querying of subscriptions for Google Play Billing Library v5 - v7.

In future diffs we will:
1. Support subscriptions in GPBL v2 - v4
2. Query for both subscription history and purchase history when IAP auto-logging is triggered.
3. Correctly parse the subscription information before sending to the server.

Note: In this diff, we also replace the ```.get(index)``` calls with ```getOrNull(index)```.  Technically,  if the index is out of range, ```get(index)``` throws an error. In our case, we were using ```AutoHandleExceptions``` to handle it, but it is better practice to just treat the error as a null value.

Reviewed By: jjiang10

Differential Revision: D62333656

fbshipit-source-id: 331a5faa3799c6b7811f6f075f34b94e6d28b2ee