Query for both subscriptions and purchases each time
Summary:
In this diff, we query for both subscription history and purchase history when IAP auto-logging is triggered.

In this diff, we update the logic to query exclusively for purchase history not purchases in these same two conditions, as purchases will always be a subset of purchase history. In the v2_v3 implementation we elected to query purchases most of the time and query purchase history at most once a day. That was because:

1. Query purchases used to be a synchronous operation, so it was preferred due to being less cpu intensive.
2. We queried on every single onActivityResume(), so cpu cost was extremely important.

Because we will only trigger the IAP logic a few times at most per app session, we can afford to just query for purchase history each time.

Likewise, we elected to use a concurrent solution in the activity lifecycle tracker to trigger the IAP logic, and not get in the way of the main/UI thread.

Reviewed By: jjiang10

Differential Revision: D62345324

fbshipit-source-id: af93a20044ceb73ce862615f86a4ddf2c089aa07