Implement improved caching logic & clear old cache
Summary:
In theory, we should never log an in-app purchase event that is older than the newest logged in-app purchase in the cache. The rationale behind this is that, if a given purchase is older than newest logged in-app purchase, because we query for all purchase history each time, then it would have already been logged whenever that *newest logged* in app purchase was. Therefore, in theory, instead of caching each purchase token for deduping, using O(n) memory, we can just cache the age of the newest logged IAP event, using O(1) memory.

There are two quirks to this:
1. When the app is first launched with this new IAP logic and there was previously logged IAP events that were cached in the old implementation, we need to first dedupe with all of those purchase events. From then on, we can just dedupe with the O(1) solution described above and clear the old cache.
2. When the app is first launched with this new IAP logic and there are no previously logged events, we can set the default "newest logged time" as 0 because we want to log all purchases no matter how old.

Reviewed By: jjiang10

Differential Revision: D62466272

fbshipit-source-id: 69c1df2ef4e51bd57a04e6348ef309061d9e3e48