Gate IAP auto-logging on launch with app_events_enabled field
Summary: Prior to this diff, IAP logging on app launch was not gated by the auto_log_app_events_enabled field, but just the implicitPurchaseLoggingEnabled field. It may be rare for an app to have implicitiPurchaseLogging Enabled but auto_log_app_events disabled, but we should still respect the auto_log_app_events field and not waste resources fetching in app purchases when auto_log_app_events is disabled.

Reviewed By: jjiang10

Differential Revision: D63271136

fbshipit-source-id: 39c333182817bb854d704e309a9e1a1f2cf95fc5