Added parameter check to implicit x manual dedupe
Summary:
As part of our IAP enhancement plan, we want to dedupe manual and implicit purchases/subscriptions that we consider likely to be referring to the same event. In previous diffs in this stack, we began to define a duplicate event by one that:

1. Occurs within the same period of time (dedupe window)
2. Is of the same value
3. Is of the same currency

To be safe, we want to also add the requirement that at least one of the other common IAP fields is equivalent. These will always be present in implicitly logged events, but may be absent in manual events.

Reviewed By: jjiang10

Differential Revision:
D64308060

Privacy Context Container: L1285343

fbshipit-source-id: 5d8197ebf0bf99c799dfb0c648fafaa81ecb2d3c