Checking if the PA manager is initialized before checking enablement
Summary: We noticed that, sometimes joining PA group request failed because PACustomAudienceClient.enabled() did not get a chance to be triggered, e.g.  PACustomAudienceClient. enabled() is called after AppEventManager starts.  In this diff, we try to enable PA when a join request happens if it is not yet initialized in order to not miss any request.

Reviewed By: youerkang

Differential Revision:
D69957967

Privacy Context Container: L1250336

fbshipit-source-id: 1409b0e8b815dc924ad9a0e252712bd538828445